fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
psycopg2-binary
passlib[bcrypt]
pydantic
python-jose

